last commit

+++ src/main/java/com/example/demo/repository/PatientRecordRepository.java
@@ -3,5 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.PatientRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PatientRecordRepository extends JpaRepository<PatientRecord, Integer> {
+public interface PatientRecordRepository extends JpaRepository<PatientRecord,Long>
+{
 }
